refactor(pwa): start decomposing the 2509-line tasks.tsx
First, behaviour-preserving slice toward breaking up the monster Tasks page.
Extract the dependency-free task-domain types + pure helpers (date math,
recurrence/sort types, reminder constants, markdown-to-plaintext) into
components/tasks/taskHelpers.ts, and move the two cleanest leaf components —
RecurrencePicker and SortableHeader — into their own files. tasks.tsx drops
~250 lines (2509 → 2262) and the shared module gives the row components a home
to import from. The large TaskRow / NewTaskRow / Tasks container remain as
follow-ups.
